(build) Getting things to work!
Before moving forward with the next release, wanted to make sure that
things were currently working.  Got access to a MicrosoftTeam instance
from Mattias, and made sure that the test.cake was working as expected.
In order to get that, had to add the missing `summary` to the example
request, otherwise there were errors returned. Also switched to
BuiltFramework.ToString() rather than TargetFramework, as this was also
causing errors.

Now that things are working, can work on updating to the different Cake
versions.